Move one more type to common
